airflow wf: close curation not working fix
