+++ workflows/requirements-test.txt
@@ -1,3 +1,5 @@
 pytest
 coverage
 pytest-cov
+pytest-vcr==1.0.2
+vcrpy==6.0.1
